requirements.txt
requests>=2.31.0
pathlib2>=2.3.7;
python-dotenv>=1.0.0
openai>=1.0.0
pinecone-client>=2.2.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
